extracted metadata exchange
